+++ components/footer-item.tsx
